+++ src/server/main.py
@@ -82,11 +82,9 @@ class Place:
         self.name = name
 
 
-@app.get("/")
-async def index(request: Request) -> Response:
-    """TODO: this is the main hompage, should have a bubble map which should
-    link ot the /model"""
-    return templates.TemplateResponse("index.html", {"request": request},)
+@app.get("/case-map")
+async def case_map(request: Request) -> Response:
+    return templates.TemplateResponse("case-map.html", {"request": request},)
 
 
 @app.get("/request-calculation")
@@ -102,7 +100,7 @@ async def request_calculation(request: Request) -> Response:
     )
 
 
-@app.get("/model")
+@app.get("/")
 async def model(request: Request, country: str = "USA") -> Response:
     """serve the main model visualization"""
     arguments = {"country": country} if country else {}
@@ -195,8 +193,7 @@ async def containment_measures(request: Request, country: str = "China") -> Resp
             ["date", "Description of measure implemented", "Source"],
         ].sort_values(by="date", ascending=False)
         sel["date"] = sel.date.dt.strftime("%Y-%m-%d")
-        args = sel.to_dict()
+        measures = [val for _, val in sel.to_dict(orient="index").items()]
     else:
-        args = None
-
-    return args
+        measures = None
+    return measures