+++ src/server/static/js/lines.js
@@ -18,20 +18,22 @@ function setGetParam(key, value) {
 let today = new Date();
 // set the dimensions and margins of the graph
 
-var chartDiv = document.getElementById("my_dataviz");
+var chartDiv = document.getElementById("my_dataviz"); //Mati: this doesn't seem to be doing anythnig
 
 // set the dimensions and margins of the graph
-var margin = { top: 10, right: 100, bottom: 30, left: 50 },
+var margin = { top: 10, right: 100, bottom: 30, left: 100 },
   width = 600,
-  height = 700;
+  height = 675;
 // append the svg object to the body of the page
+var graphSize = "75%";
 var svg = d3
   .select("#my_dataviz")
+  .style("width", graphSize)
+  .style("padding-bottom", graphSize)
   .append("svg")
   .attr("preserveAspectRatio", "xMinYMin meet")
   .attr("viewBox", "0 0 750 750")
   .classed("svg-content", true)
-
   .append("g")
   .attr("transform", "translate(" + margin.left + "," + margin.top + ")");
 
@@ -68,6 +70,21 @@ function getSelectedCountry(data) {
   return c && data.includes(c) ? c : "China";
 }
 
+function getMaxYValueForCountry(countryBetas, selectedCountry) {
+  var highestVals = [];
+  countryBetas.forEach(b => {
+    if (b.country !== selectedCountry) {
+      return;
+    }
+    b.items.forEach(i => {
+      a = [...i];
+      a.shift(); //// remove date
+      highestVals.push(Math.max(...a));
+    });
+  });
+  return Math.max(...highestVals);
+}
+
 //Read the data
 d3.csv(
   "https://storage.googleapis.com/static-covid/static/line-data-v2.csv?cache_bump=2"
@@ -77,7 +94,7 @@ d3.csv(
   var allGroup = getCountries(data);
   countryBetas = getCountryBetaData(data);
   selectedCountry = getSelectedCountry(allGroup);
-  //console.log(allGroup)
+  console.log(countryBetas, "betas");
   selectedCountryBeta = countryBetas.find(
     r => r.country === selectedCountry && r.beta === selectedBeta
   );
@@ -94,7 +111,8 @@ d3.csv(
     }) // text showed in the menu
     .attr("value", function(d) {
       return d;
-    }); // corresponding value returned by the button
+    })
+    .sort(); // corresponding value returned by the button
 
   var xDomain = d3.extent(selectedCountryBeta.items, function(d) {
     return d[0];
@@ -103,20 +121,17 @@ d3.csv(
   var x = d3
     .scaleTime()
     .range([0, width])
-    // .domain( [0,1000])
     .domain(xDomain);
-  // .ticks(d3.time.months, 1).tickFormat(d3.time.format("%b"));;
   svg
     .append("g")
+    .style("font-size", "20px")
     .attr("transform", "translate(0," + height + ")")
     .call(
       d3
         .axisBottom(x)
         .ticks(6)
-        .tickFormat(d3.timeFormat("%Y-%m-%d"))
+        .tickFormat(d3.timeFormat("%b %Y"))
     );
-
-  var yDomain = [0, 200 / 1000];
 
   // text label for the x axis
   svg
@@ -125,10 +140,12 @@ d3.csv(
     .style("fill", "#a9a9ac")
     .attr(
       "transform",
-      "translate(" + width / 2 + " ," + (height + margin.top + 20) + ")"
+      "translate(" + width / 2 + " ," + (height + margin.top + 40) + ")"
     )
     .style("text-anchor", "middle")
     .text("Date");
+
+  var yDomain = [0, getMaxYValueForCountry(countryBetas, selectedCountry)];
 
   // Add Y axis
   var y = d3
@@ -136,7 +153,15 @@ d3.csv(
     .domain(yDomain)
     //.domain([0, d3.max(selectedCountryBeta.items, function(d) { return Math.max(+d[1], +d[2], +d[3], +d[4]); })])
     .range([height, 0]);
-  svg.append("g").call(d3.axisLeft(y));
+  var yAxis = svg
+    .append("g")
+    .style("font-size", "20px")
+    .call(
+      d3
+        .axisLeft(y)
+        .ticks(10)
+        .tickFormat(d3.format(".0%"))
+    );
 
   // text label for the y axis
   svg
@@ -148,7 +173,10 @@ d3.csv(
     .attr("x", 0 - height / 2)
     .attr("dy", "1em")
     .style("text-anchor", "middle")
-    .text("Value");
+    .text("Active infections");
+
+  // Set font size for axis labels
+  svg.style("font-size", "22px");
 
   function drawLine(i, color) {
     return svg
@@ -278,6 +306,7 @@ d3.csv(
   d3.select("#selectButton").property("value", selectedCountry);
   // update the containment measures with the new selected country
   update_containment_measures(selectedCountry);
+  update_country_in_text(selectedCountry);
 
   // A function that update the chart
   function update({
@@ -287,6 +316,8 @@ d3.csv(
     selectedCountryBeta = countryBetas.find(
       r => r.country === country && r.beta === beta
     );
+    yDomain[1] = getMaxYValueForCountry(countryBetas, selectedCountry);
+    y = y.domain(yDomain);
     function updateLine(myLine, i) {
       myLine
         .datum(selectedCountryBeta.items)
@@ -305,24 +336,35 @@ d3.csv(
         );
     }
 
+    yAxis
+      .transition()
+      .duration(1000)
+      .call(d3.axisLeft(y));
+
     updateLine(line1, 1);
     updateLine(line2, 2);
     updateLine(line3, 3);
     updateLine(line4, 4);
     updateLine(line5, 5);
     updateLine(line6, 6);
+
+    // svg.select(".y.axis") // change the y axis
+    // .duration(750)
+    // .call(y);
   }
 
   // When the button is changed, run the updateChart function
   d3.select("#selectButton").on("change", function(d) {
     // recover the option that has been chosen
-    var selectedOption = d3.select(this).property("value");
+    selectedCountry = d3.select(this).property("value");
     // change url param
-    setGetParam("selection", selectedOption);
+    setGetParam("selection", selectedCountry);
     // run the updateChart function with this selected option
-    update({ country: selectedOption });
+    update({ country: selectedCountry });
     // update the containment measures with the new selected country
-    update_containment_measures(selectedOption);
+    update_containment_measures(selectedCountry);
+    // update the name of the country in the text below the graph
+    update_country_in_text(selectedCountry);
   });
 
   d3.select(".beta-0").on("click", function() {
@@ -340,6 +382,13 @@ d3.csv(
 
   //console.log("RUNNING D3");
 });
+
+function update_country_in_text(selectedCountry) {
+  var countrySpans = jQuery(".selected-country");
+  for (i = 0; i < countrySpans.length; i++) {
+    countrySpans[i].innerHTML = selectedCountry;
+  }
+}
 
 function containment_entry(date = "", text = "", source_link = "") {
   /* write that jinja code with js for model.html template sidebar with containment measures entry
@@ -375,18 +424,26 @@ function update_containment_measures(selectedOption) {
       containmentMeasuresDiv.textContent = "";
       var divTitle = document.createElement("H5");
       divTitle.innerHTML = "Containment measures";
+
       containmentMeasuresDiv.append(divTitle);
+      var containmentMeasuresSource = document.createElement("a");
+      var linkText = document.createTextNode("(data source)");
+      containmentMeasuresSource.appendChild(linkText);
+      containmentMeasuresSource.title = "(data source)";
+      containmentMeasuresSource.href =
+        "https://www.notion.so/977d5e5be0434bf996704ec361ad621d?v=aa8e0c75520a479ea48f56cb4c289b7e";
+      containmentMeasuresDiv.append(containmentMeasuresSource);
+
       if (data != undefined) {
-        // format each entry and append it to the containmentMeasuresDiv
-        for (let i in data["Description of measure implemented"]) {
+        data.forEach(function(item, index) {
           containmentMeasuresDiv.appendChild(
             containment_entry(
-              (date = data["date"][i]),
-              (text = data["Description of measure implemented"][i]),
-              (source_link = data["Source"][i])
+              (date = item["date"]),
+              (text = item["Description of measure implemented"]),
+              (source_link = item["Source"])
             )
           );
-        }
+        });
       } else {
         var emptyDatasetMsg = document.createElement("P");
         emptyDatasetMsg.innerHTML =